增加线缆渲染
完善对面的支持
增加创造模式Tab

+++ SimElectricity/src/main/java/simElectricity/API/IBaseComponent.java
@@ -7,7 +7,7 @@ public interface IBaseComponent {
 	 * sinks energy For a energy source, the smaller the resistance is, the
 	 * lower voltage drop it will have, this means it can drive heavier load
 	 */
-	int getResistance();
+	float getResistance();
 
 	/**
 	 * For load: Return the max power the load can sink, 0 for infinite For
